fix conflicting css and move avatar components into own folder
